+++ client/src/components/AdminLogin/index.js
@@ -1,0 +1,1 @@
+export { default } from './AdminLogin';
