+++ client/src/pages/admin/index.module.scss
@@ -1,0 +1,1 @@
+@import '../../styles/colors';
